regListToMnemonic() takes uint16 instead of uint8

+++ hardware/arm/thumb2_32bit.go
@@ -3259,7 +3259,7 @@ func (arm *ARM) decode32bitThumb2LoadStoreMultiple(opcode uint16) decodeFunction
 						return &DisasmEntry{
 							Is32bit:  true,
 							Operator: "POP",
-							Operand:  fmt.Sprintf("{%s}", reglistToMnemonic('R', uint8(regList), "")),
+							Operand:  fmt.Sprintf("{%s}", reglistToMnemonic('R', regList, "")),
 						}
 					}
 
@@ -3301,7 +3301,7 @@ func (arm *ARM) decode32bitThumb2LoadStoreMultiple(opcode uint16) decodeFunction
 						return &DisasmEntry{
 							Is32bit:  true,
 							Operator: "LDMIA",
-							Operand:  fmt.Sprintf("R%d%c, {%s}", Rn, writebackSign, reglistToMnemonic('R', uint8(regList), "")),
+							Operand:  fmt.Sprintf("R%d%c, {%s}", Rn, writebackSign, reglistToMnemonic('R', regList, "")),
 						}
 					}
 
@@ -3350,7 +3350,7 @@ func (arm *ARM) decode32bitThumb2LoadStoreMultiple(opcode uint16) decodeFunction
 					return &DisasmEntry{
 						Is32bit:  true,
 						Operator: "STMIA",
-						Operand:  fmt.Sprintf("R%d%c, {%s}", Rn, writebackSign, reglistToMnemonic('R', uint8(regList), "")),
+						Operand:  fmt.Sprintf("R%d%c, {%s}", Rn, writebackSign, reglistToMnemonic('R', regList, "")),
 					}
 				}
 
@@ -3398,7 +3398,7 @@ func (arm *ARM) decode32bitThumb2LoadStoreMultiple(opcode uint16) decodeFunction
 					return &DisasmEntry{
 						Is32bit:  true,
 						Operator: "LDMDB",
-						Operand:  fmt.Sprintf("R%d%c, {%s}", Rn, writebackSign, reglistToMnemonic('R', uint8(regList), "")),
+						Operand:  fmt.Sprintf("R%d%c, {%s}", Rn, writebackSign, reglistToMnemonic('R', regList, "")),
 					}
 				}
 
@@ -3445,7 +3445,7 @@ func (arm *ARM) decode32bitThumb2LoadStoreMultiple(opcode uint16) decodeFunction
 						return &DisasmEntry{
 							Is32bit:  true,
 							Operator: "PUSH",
-							Operand:  fmt.Sprintf("{%s}", reglistToMnemonic('R', uint8(regList), "")),
+							Operand:  fmt.Sprintf("{%s}", reglistToMnemonic('R', regList, "")),
 						}
 					}
 
@@ -3482,7 +3482,7 @@ func (arm *ARM) decode32bitThumb2LoadStoreMultiple(opcode uint16) decodeFunction
 						return &DisasmEntry{
 							Is32bit:  true,
 							Operator: "STMDB",
-							Operand:  fmt.Sprintf("R%d%c, {%s}", Rn, writebackSign, reglistToMnemonic('R', uint8(regList), "")),
+							Operand:  fmt.Sprintf("R%d%c, {%s}", Rn, writebackSign, reglistToMnemonic('R', regList, "")),
 						}
 					}
